Updated post details page
Instead of sending a router state with post details url, a router with params was created. The PageDetailComponent receives post id and call post service to get data from server. HttpInterceptor intercepts 404 - Not found error and redirects to NotFound component, if post with specified id was not found. Some new icons for working with particular post were added.